requirements.txt
cssutils
csscompressor
pillow